leave

+++ src/main/java/tr/com/bacompany/bacrm/service/impl/TimesheetServiceImpl.java
@@ -4,12 +4,15 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tr.com.bacompany.bacrm.converter.TimesheetConverter;
-import tr.com.bacompany.bacrm.data.dto.TimesheetDto;
-import tr.com.bacompany.bacrm.data.entity.Timesheet;
+import tr.com.bacompany.bacrm.data.dto.timesheet.TimesheetDto;
+import tr.com.bacompany.bacrm.data.entity.timesheet.EnumTimesheetStatus;
+import tr.com.bacompany.bacrm.data.entity.timesheet.Timesheet;
 import tr.com.bacompany.bacrm.data.exception.ResourceNotFoundException;
 import tr.com.bacompany.bacrm.repository.TimesheetRepository;
 import tr.com.bacompany.bacrm.service.TimesheetService;
 
+import javax.transaction.Transactional;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -25,9 +28,27 @@ public class TimesheetServiceImpl implements TimesheetService {
     }
 
     @Override
-    public TimesheetDto add(TimesheetDto timesheetDto) {
-        Timesheet timesheet = timesheetRepository.save(TimesheetConverter.toEntity(timesheetDto));
-        return TimesheetConverter.toDto(timesheet);
+    public List<TimesheetDto> saveTimesheet(List<TimesheetDto> timesheetDtoList) {
+        timesheetDtoList.forEach(e -> e.setStatus(EnumTimesheetStatus.SAVED));
+        return this.save(timesheetDtoList);
+    }
+
+    @Override
+    public List<TimesheetDto> approveTimesheetByUser(List<TimesheetDto> timesheetDtoList) {
+        timesheetDtoList.forEach(e -> e.setStatus(EnumTimesheetStatus.APPROVED_BY_USER));
+        return this.save(timesheetDtoList);
+    }
+
+    @Override
+    public List<TimesheetDto> approveTimesheetByManager(List<TimesheetDto> timesheetDtoList) {
+        timesheetDtoList.forEach(e -> e.setStatus(EnumTimesheetStatus.APPROVED_BY_MANAGER));
+        return this.save(timesheetDtoList);
+    }
+
+    @Override
+    public List<TimesheetDto> rejectTimesheetByManager(List<TimesheetDto> timesheetDtoList) {
+        timesheetDtoList.forEach(e -> e.setStatus(EnumTimesheetStatus.REJECTED));
+        return this.save(timesheetDtoList);
     }
 
     @Override
@@ -46,20 +67,40 @@ public class TimesheetServiceImpl implements TimesheetService {
     }
 
     @Override
-    public TimesheetDto update(TimesheetDto workDto) throws ResourceNotFoundException {
-        this.getBy(workDto.getId());
-        Timesheet newTimesheet = TimesheetConverter.toEntity(workDto);
-        return TimesheetConverter.toDto(timesheetRepository.save(newTimesheet));
+    public List<TimesheetDto> update(List<TimesheetDto> timesheetDtoList) throws ResourceNotFoundException {
+        List<TimesheetDto> resultTimesheetDtoList = new ArrayList<>();
+        for (TimesheetDto timesheetDto : timesheetDtoList) {
+            this.getBy(timesheetDto.getId());
+            Timesheet newTimesheet = TimesheetConverter.toEntity(timesheetDto);
+            TimesheetDto newTimesheetDto = TimesheetConverter.toDto(timesheetRepository.save(newTimesheet));
+            resultTimesheetDtoList.add(newTimesheetDto);
+        }
+        return resultTimesheetDtoList;
     }
 
     @Override
-    public boolean delete(TimesheetDto timesheetDto) {
+    public boolean delete(TimesheetDto timesheetDto) throws ResourceNotFoundException {
         try {
-            Timesheet timesheet = TimesheetConverter.toEntity(timesheetDto);
+            Optional<Timesheet> optTimesheet = timesheetRepository.findById(timesheetDto.getId());
+            if (!optTimesheet.isPresent()) {
+                throw new ResourceNotFoundException("Timesheet is not found.", "Timesheet");
+            }
+            Timesheet timesheet = optTimesheet.get();
             timesheetRepository.delete(timesheet);
         } catch (Exception ex) {
             return false;
         }
         return true;
+    }
+
+    @Transactional
+    private List<TimesheetDto> save(List<TimesheetDto> timesheetDtoList) {
+        List<TimesheetDto> resultTimesheetDtoList = new ArrayList<>();
+        for (TimesheetDto timesheetDto : timesheetDtoList) {
+            Timesheet timesheet = timesheetRepository.save(TimesheetConverter.toEntity(timesheetDto));
+            TimesheetDto savedTimesheet = TimesheetConverter.toDto(timesheet);
+            resultTimesheetDtoList.add(savedTimesheet);
+        }
+        return resultTimesheetDtoList;
     }
 }
